refactor: move CustomDialog for Seester to SemesterList component
