+++ frontend/public/assets/i18n/translations/en/formElements.json
@@ -125,5 +125,5 @@
   "create_account_label": "Create an Account",
   "forgot_password_label": "Forgot password?",
   "dont_have_an_account": "Don't have an account?",
-  "reset_password_helper_text": "Just enter your email address below and we'll send you a link to reset your password!"
+  "reset_password_helper_text": "Just enter your email address below and we'll send you a link to reset your password!",
 }
