+++ frontend/src/components/Semester/SemesterList.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { FaEdit, FaUsers, FaFileArchive } from 'react-icons/fa';
 import { MdDelete, MdDonutSmall } from 'react-icons/md';
 import { useTranslation } from 'react-i18next';
@@ -7,7 +7,7 @@ import { GiSightDisabled, IoMdEye, FaCopy } from 'react-icons/all';
 import { get } from 'lodash';
 import Card from '../../share/Card/Card';
 import NotFound from '../../share/NotFound/NotFound';
-import { dialogTypes } from '../../constants/dialogs';
+import { dialogTypes, dialogCloseButton } from '../../constants/dialogs';
 import {
     EDIT_TITLE,
     DELETE_TITLE,
@@ -15,6 +15,7 @@ import {
     SEMESTERY_LABEL,
     FORM_SHOW_GROUPS,
     SET_DEFAULT_TITLE,
+    SEMESTER_COPY_LABEL,
 } from '../../constants/translationLabels/formElements';
 import {
     COMMON_SET_DISABLED,
@@ -26,6 +27,8 @@ import {
 } from '../../constants/translationLabels/common';
 import { search } from '../../helper/search';
 import { getGroupsOptionsForSelect } from '../../utils/selectUtils';
+import SemesterCopyForm from '../../containers/SemesterPage/SemesterCopyForm';
+import CustomDialog from '../../containers/Dialogs/CustomDialog';
 
 const SemesterList = (props) => {
     const { t } = useTranslation('formElements');
@@ -33,18 +36,24 @@ const SemesterList = (props) => {
         archived,
         disabled,
         setSemesterId,
-        setConfirmDialogType,
-        setOpenSubDialog,
-        setIsOpenSemesterCopyForm,
         setOpenGroupsDialog,
         term,
         semesters,
         setSemesterOptions,
         selectSemester,
         createArchivedSemester,
+        setOpenConfirmDialog,
+        updateSemester,
+        removeSemesterCard,
+        setDefaultSemesterById,
+        isOpenConfirmDialog,
+        semesterCopy,
+        semesterId,
         // archivedSemesters,
         // getArchivedSemester,
     } = props;
+    const [isOpenSemesterCopyForm, setIsOpenSemesterCopyForm] = useState(false);
+    const [confirmDialogType, setConfirmDialogType] = useState('');
 
     const searchArr = ['year', 'description', 'startDay', 'endDay'];
 
@@ -53,82 +62,144 @@ const SemesterList = (props) => {
     const showConfirmDialog = (id, dialogType) => {
         setSemesterId(id);
         setConfirmDialogType(dialogType);
-        setOpenSubDialog(true);
+        setOpenConfirmDialog(true);
+    };
+    const submitSemesterCopy = ({ toSemesterId }) => {
+        semesterCopy({
+            fromSemesterId: semesterId,
+            toSemesterId,
+        });
+        setIsOpenSemesterCopyForm(false);
+    };
+    const closeSemesterCopyForm = () => {
+        setIsOpenSemesterCopyForm(false);
+    };
+    const changeSemesterDisabledStatus = (currentSemesterId) => {
+        const foundSemester = semesters.find(
+            (semesterItem) => semesterItem.id === currentSemesterId,
+        );
+        const changeDisabledStatus = {
+            [dialogTypes.SET_VISIBILITY_ENABLED]: updateSemester({
+                ...foundSemester,
+                disable: false,
+            }),
+            [dialogTypes.SET_VISIBILITY_DISABLED]: updateSemester({
+                ...foundSemester,
+                disable: true,
+            }),
+        };
+        return changeDisabledStatus[confirmDialogType];
+    };
+
+    const acceptConfirmDialog = (currentSemesterId) => {
+        setOpenConfirmDialog(false);
+        const isDisabled = disabled;
+        if (confirmDialogType === dialogTypes.SET_DEFAULT) {
+            setDefaultSemesterById(currentSemesterId, isDisabled);
+        } else if (confirmDialogType !== dialogTypes.DELETE_CONFIRM) {
+            changeSemesterDisabledStatus(currentSemesterId);
+        } else {
+            removeSemesterCard(currentSemesterId);
+        }
     };
     // it doesnt work, need to finish implement archeved functionality
     // const handleSemesterArchivedPreview = (currentSemesterId) => {
     //     getArchivedSemester(+currentSemesterId);
     // };
     return (
-        <section className="container-flex-wrap wrapper">
-            {visibleItems.length === 0 && <NotFound name={t(SEMESTERY_LABEL)} />}
-            {visibleItems.map((semesterItem) => {
-                const semDays = semesterItem.semester_days.map((day) =>
-                    t(`common:day_of_week_${day}`),
-                );
-                const groups = get(semesterItem, 'semester_groups')
-                    ? getGroupsOptionsForSelect(semesterItem.semester_groups)
-                    : [];
+        <>
+            {isOpenConfirmDialog && (
+                <CustomDialog
+                    type={confirmDialogType}
+                    whatDelete="semester"
+                    open={isOpenConfirmDialog}
+                    handelConfirm={() => acceptConfirmDialog(semesterId)}
+                />
+            )}
+            {isOpenSemesterCopyForm && (
+                <CustomDialog
+                    title={t(SEMESTER_COPY_LABEL)}
+                    open={isOpenSemesterCopyForm}
+                    onClose={closeSemesterCopyForm}
+                    buttons={[dialogCloseButton(closeSemesterCopyForm)]}
+                >
+                    <SemesterCopyForm
+                        semesterId={semesterId}
+                        onSubmit={submitSemesterCopy}
+                        submitButtonLabel={t(COPY_LABEL)}
+                        semesters={semesters}
+                    />
+                </CustomDialog>
+            )}
+            <section className="container-flex-wrap wrapper">
+                {visibleItems.length === 0 && <NotFound name={t(SEMESTERY_LABEL)} />}
+                {visibleItems.map((semesterItem) => {
+                    const semDays = semesterItem.semester_days.map((day) =>
+                        t(`common:day_of_week_${day}`),
+                    );
+                    const groups = get(semesterItem, 'semester_groups')
+                        ? getGroupsOptionsForSelect(semesterItem.semester_groups)
+                        : [];
 
-                return (
-                    <Card
-                        key={semesterItem.id}
-                        additionClassName={`semester-card done-card ${
-                            semesterItem.currentSemester ? 'current' : ''
-                        }`}
-                    >
-                        <div className="cards-btns">
-                            {!(disabled || archived) && (
-                                <>
-                                    <IoMdEye
+                    return (
+                        <Card
+                            key={semesterItem.id}
+                            additionClassName={`semester-card done-card ${
+                                semesterItem.currentSemester ? 'current' : ''
+                            }`}
+                        >
+                            <div className="cards-btns">
+                                {!(disabled || archived) && (
+                                    <>
+                                        <IoMdEye
+                                            className="svg-btn copy-btn"
+                                            title={t(COMMON_SET_DISABLED)}
+                                            onClick={() => {
+                                                showConfirmDialog(
+                                                    semesterItem.id,
+                                                    dialogTypes.SET_VISIBILITY_DISABLED,
+                                                );
+                                            }}
+                                        />
+                                        <FaEdit
+                                            className="svg-btn edit-btn"
+                                            title={t(EDIT_TITLE)}
+                                            onClick={() => {
+                                                selectSemester(semesterItem.id);
+                                            }}
+                                        />
+                                        <FaCopy
+                                            className="svg-btn copy-btn"
+                                            title={t(COPY_LABEL)}
+                                            onClick={() => {
+                                                setIsOpenSemesterCopyForm(true);
+                                                setSemesterId(semesterItem.id);
+                                            }}
+                                        />
+                                        {!semesterItem.currentSemester && (
+                                            <FaFileArchive
+                                                className="svg-btn archive-btn"
+                                                title={t(COMMON_MAKE_ARCHIVE)}
+                                                onClick={() => {
+                                                    createArchivedSemester(semesterItem.id);
+                                                }}
+                                            />
+                                        )}
+                                    </>
+                                )}
+                                {disabled && !archived && (
+                                    <GiSightDisabled
                                         className="svg-btn copy-btn"
-                                        title={t(COMMON_SET_DISABLED)}
+                                        title={t(COMMON_SET_ENABLED)}
                                         onClick={() => {
                                             showConfirmDialog(
                                                 semesterItem.id,
-                                                dialogTypes.SET_VISIBILITY_DISABLED,
+                                                dialogTypes.SET_VISIBILITY_ENABLED,
                                             );
                                         }}
                                     />
-                                    <FaEdit
-                                        className="svg-btn edit-btn"
-                                        title={t(EDIT_TITLE)}
-                                        onClick={() => {
-                                            selectSemester(semesterItem.id);
-                                        }}
-                                    />
-                                    <FaCopy
-                                        className="svg-btn copy-btn"
-                                        title={t(COPY_LABEL)}
-                                        onClick={() => {
-                                            setIsOpenSemesterCopyForm(true);
-                                            setSemesterId(semesterItem.id);
-                                        }}
-                                    />
-                                    {!semesterItem.currentSemester && (
-                                        <FaFileArchive
-                                            className="svg-btn archive-btn"
-                                            title={t(COMMON_MAKE_ARCHIVE)}
-                                            onClick={() => {
-                                                createArchivedSemester(semesterItem.id);
-                                            }}
-                                        />
-                                    )}
-                                </>
-                            )}
-                            {disabled && !archived && (
-                                <GiSightDisabled
-                                    className="svg-btn copy-btn"
-                                    title={t(COMMON_SET_ENABLED)}
-                                    onClick={() => {
-                                        showConfirmDialog(
-                                            semesterItem.id,
-                                            dialogTypes.SET_VISIBILITY_ENABLED,
-                                        );
-                                    }}
-                                />
-                            )}
-                            {/* {archived && (
+                                )}
+                                {/* {archived && (
                                 <IoMdEye
                                     className="svg-btn copy-btn"
                                     title={t(COMMON_PREVIEW)}
@@ -137,61 +208,65 @@ const SemesterList = (props) => {
                                     }}
                                 />
                             )} */}
-                            <MdDelete
-                                className="svg-btn delete-btn"
-                                title={t(DELETE_TITLE)}
-                                onClick={() =>
-                                    showConfirmDialog(semesterItem.id, dialogTypes.DELETE_CONFIRM)
-                                }
+                                <MdDelete
+                                    className="svg-btn delete-btn"
+                                    title={t(DELETE_TITLE)}
+                                    onClick={() =>
+                                        showConfirmDialog(
+                                            semesterItem.id,
+                                            dialogTypes.DELETE_CONFIRM,
+                                        )
+                                    }
+                                />
+
+                                <MdDonutSmall
+                                    className={`svg-btn edit-btn ${
+                                        semesterItem.defaultSemester ? 'default' : ''
+                                    }`}
+                                    title={t(SET_DEFAULT_TITLE)}
+                                    onClick={() =>
+                                        showConfirmDialog(semesterItem.id, dialogTypes.SET_DEFAULT)
+                                    }
+                                />
+                            </div>
+
+                            <p className="semester-card__description">
+                                <small>{`${t(SEMESTER_LABEL)}:`}</small>
+                                <b>{semesterItem.description}</b>
+                                {` ( ${semesterItem.year} )`}
+                            </p>
+                            <p className="semester-card__description">
+                                <b>
+                                    {semesterItem.startDay} - {semesterItem.endDay}
+                                </b>
+                            </p>
+                            <p className="semester-card__description">
+                                {`${t(COMMON_DAYS_LABEL)}: `}
+                                {semDays.join(', ')}
+                            </p>
+                            <p className="semester-card__description">
+                                {`${t(COMMON_CLASS_SCHEDULE_MANAGEMENT_TITLE)}: `}
+                                {semesterItem.semester_classes
+                                    .map((classItem) => {
+                                        return classItem.class_name;
+                                    })
+                                    .join(', ')}
+                            </p>
+
+                            <FaUsers
+                                title={t(FORM_SHOW_GROUPS)}
+                                className="svg-btn copy-btn  semester-groups"
+                                onClick={() => {
+                                    setSemesterId(semesterItem.id);
+                                    setSemesterOptions(groups);
+                                    setOpenGroupsDialog(true);
+                                }}
                             />
-
-                            <MdDonutSmall
-                                className={`svg-btn edit-btn ${
-                                    semesterItem.defaultSemester ? 'default' : ''
-                                }`}
-                                title={t(SET_DEFAULT_TITLE)}
-                                onClick={() =>
-                                    showConfirmDialog(semesterItem.id, dialogTypes.SET_DEFAULT)
-                                }
-                            />
-                        </div>
-
-                        <p className="semester-card__description">
-                            <small>{`${t(SEMESTER_LABEL)}:`}</small>
-                            <b>{semesterItem.description}</b>
-                            {` ( ${semesterItem.year} )`}
-                        </p>
-                        <p className="semester-card__description">
-                            <b>
-                                {semesterItem.startDay} - {semesterItem.endDay}
-                            </b>
-                        </p>
-                        <p className="semester-card__description">
-                            {`${t(COMMON_DAYS_LABEL)}: `}
-                            {semDays.join(', ')}
-                        </p>
-                        <p className="semester-card__description">
-                            {`${t(COMMON_CLASS_SCHEDULE_MANAGEMENT_TITLE)}: `}
-                            {semesterItem.semester_classes
-                                .map((classItem) => {
-                                    return classItem.class_name;
-                                })
-                                .join(', ')}
-                        </p>
-
-                        <FaUsers
-                            title={t(FORM_SHOW_GROUPS)}
-                            className="svg-btn copy-btn  semester-groups"
-                            onClick={() => {
-                                setSemesterId(semesterItem.id);
-                                setSemesterOptions(groups);
-                                setOpenGroupsDialog(true);
-                            }}
-                        />
-                    </Card>
-                );
-            })}
-        </section>
+                        </Card>
+                    );
+                })}
+            </section>
+        </>
     );
 };
